requirements.txt
Django==4.2.13
djangorestframework==3.15.2
django-cors-headers==4.3.1
celery==5.3.6
redis==5.0.3
Pillow>=11.0.0
opencv-python-headless>=4.10.0
numpy>=2.0.0
scipy>=1.13.0
django-storages==1.14.2
python-dotenv==1.0.1
# psycopg2-binary requires MSVC build tools on Windows.
# For PostgreSQL support, install it separately or use Linux/WSL.
# For development, SQLite is used by default.
